push current api

+++ app/Api/V1/Controllers/AWSController.php
@@ -112,18 +112,46 @@ class AWSController extends Controller
         $jsonData = json_decode($tempJson);
         $jsonData = get_object_vars($jsonData);
         //return $jsonData;
+        $creds = [
+            'credentials' => [
+                'key'    => 'AKIAJ7MQ4FVAL5MNQZ6A',
+                'secret' => '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',
+                ],
+            'version' => 'latest',
+            'region' => $jsonData['region_id'],
+         ];
+        
+        /*$provider = CredentialProvider::instanceProfile();
+        $memoizedProvider = CredentialProvider::memoize($provider);
+        $creds = [
+                    'credentials' => $memoizedProvider,
+                    'version' => 'latest',
+                    'region' => $jsonData['region_id'],
+                 ];*/
         $cloud = new CloudHelper();
-        $result = $cloud->call_cloud_api_command($jsonData);
+        $result = $cloud->call_cloud_api_command($jsonData,$creds);
         return $result;         
     }
 
     public function addFeature($feature_file, $feature_name)
     {
         $result = [ 'error_code' => '0', 'error_msg' => '', 'result' => []];
-        $body = $this->customFeatureTemplate();
+        $body = $this->customFeatureTemplate($feature_name);
         $res = file_put_contents($feature_file,$body);
         if(!$res) {
             $result = [ 'error_code' => 'XCFeatureFileWriteError', 'error_msg' => 'ERROR: Could not save feature code.', 'result' => ['body' => $body]];
+        }
+        $res = mkdir("/var/www/xcloudserverapi/resources/views/features/{$feature_name}");
+        if(!$res) {
+            $result = [ 'error_code' => 'XCFeatureFileWriteError', 'error_msg' => 'ERROR: Could not create frontend directory.', 'result' => ['body' => $body]];
+        }
+        $res = mkdir("/var/www/xcloudserverapi/resources/views/features/{$feature_name}/html");
+        if(!$res) {
+            $result = [ 'error_code' => 'XCFeatureFileWriteError', 'error_msg' => 'ERROR: Could not create frontend directory.', 'result' => ['body' => $body]];
+        }
+        $res = mkdir("/var/www/xcloudserverapi/resources/views/features/{$feature_name}/js");
+        if(!$res) {
+            $result = [ 'error_code' => 'XCFeatureFileWriteError', 'error_msg' => 'ERROR: Could not create frontend directory.', 'result' => ['body' => $body]];
         }
         return $result;
     }
@@ -164,15 +192,43 @@ class AWSController extends Controller
         return $result;
     }
 
+    public function rrmdir($dir){ 
+        if (is_dir($dir)) { 
+            $objects = scandir($dir); 
+            foreach ($objects as $object) { 
+                if ($object != "." && $object != "..") { 
+                    if (is_dir($dir."/".$object)){
+                        rrmdir($dir."/".$object);
+                    }
+                    else{
+                        unlink($dir."/".$object); 
+                    }
+                } 
+            }
+            rmdir($dir); 
+        } 
+    }
 
-    public function removeFeature($feature_file)
+
+    public function removeFeature($feature_file, $feature_name)
     {
         $result = [ 'error_code' => '0', 'error_msg' => '', 'result' => []];
+        if(!file_exists($feature_file)) {
+            return $result;
+        }
         $res = unlink($feature_file);
         if(!$res)
         {
             $result = [ 'error_code' => 'XCFeatureFileDeleteError', 'error_msg' => 'ERROR: Could not delete feature.', 'result' => []];
         }
+        $frontend_dir = "/var/www/xcloudserverapi/resources/views/features/{$feature_name}";
+        $output = shell_exec("rm -r {$frontend_dir}");
+        // $this->rrmdir($frontend_dir);
+        // $res = rmdir("/var/www/xcloudserverapi/resources/views/features/{$feature_name}");
+        // if(!$res)
+        // {
+        //     $result = [ 'error_code' => 'XCFeatureFileDeleteError', 'error_msg' => 'ERROR: Could not delete feature frontend directory.', 'result' => []];
+        // }
         return $result;
     }
 
@@ -304,10 +360,10 @@ Class test {
         $result = [];
         $feature_file = '';
         if(isset($param['feature'])) {
-            $feature_file = "/var/www/xcloudclientapi/app/FeatureCommands/{$param['feature']}.php";
+            $feature_file = "/var/www/xcloudserverapi/app/FeatureCommands/{$param['feature']}.php";
         }
         
-        $settings_file = "/var/www/xcloudclientapi/.env";
+        $settings_file = "/var/www/xcloudserverapi/.env";
         switch($param['command']) {
             case 'get' : 
                 $result = $this->getFeature($feature_file);
@@ -319,7 +375,7 @@ Class test {
                 $result = $this->saveFeature($feature_file, $param['feature'], $param['body']);
                 break;
             case 'remove_feature' : 
-                $result = $this->removeFeature($feature_file);
+                $result = $this->removeFeature($feature_file, $param['feature']);
                 break;
             case 'add_function' : 
                 $result = $this->addFunction($feature_file, $param['feature'], $param['function_name']);
@@ -352,15 +408,18 @@ Class test {
         $param = $request->all();
         $awsclass = $param['awsclass'];
         $awscommand = $param['awscommand'];
-        //$awskey = $param['awskey'];
-        //$awssecret = $param['awssecret'];
+        $awskey = $param['awskey'];
+        $awssecret = $param['awssecret'];
         $awsregion = $param['awsregion'];
         $client = [];
 
-        $provider = CredentialProvider::instanceProfile();
-        $memoizedProvider = CredentialProvider::memoize($provider);
+        // $provider = CredentialProvider::instanceProfile();
+        // $memoizedProvider = CredentialProvider::memoize($provider);
         $creds = [
-                    'credentials' => $memoizedProvider,
+                    'credentials' => [
+                        'key'    => $awskey,
+                        'secret' => $awssecret,
+                        ],
                     'version' => 'latest',
                     'region' => $awsregion,
                  ];
@@ -430,6 +489,7 @@ Class test {
     {
         $str = "<?php
 namespace App\FeatureCommands;
+use App\CloudHelper;
 use DB;
 class {$feature}
 {
@@ -490,6 +550,110 @@ class {$feature}
     public function scriptLang(Request $request)
     {
         $result = ['error_code' => '0', 'error_msg' => 'SUCCESS:', 'result' => ['language' => 'php'] ];
+    }
+
+    public function frontendEdit(Request $request)
+    {
+        $user = JWTAuth::parseToken()->authenticate();
+        $param = $request->all();
+        $error_msg = 'Success';
+        $error_code = '0';
+        $result = [];
+        $feature_file = '';
+        switch($param['command']) {
+            case 'create' : 
+                $result = $this->createFrontendFile($param['feature'], $param['filename'], $param['editmode']);
+                break;
+            case 'getfiles' : 
+                $result = $this->readDirectory($param['feature'], $param['editmode']);
+                break;
+            case 'savecode' : 
+                $result = $this->saveCodeFile($param);
+                break;
+            case 'getfilecode' : 
+                $result = $this->getFileCode($param);
+                break;
+            default:
+                $result = ['error_code' => 'XCCommandNotFoundError', 'error_msg' => 'ERROR: Unknown command.', 'result' => [] ];
+        }
+        return $result;
+    }
+
+    public function createFrontendFile($feature_name, $filename, $editmode)
+    {
+        $result = ['error_code' => 'FILEWRITEERROR', 'error_msg' => 'Unable to write file', 'result' => [] ];
+        if($editmode == "html")
+        {
+            $filepath = "/var/www/xcloudserverapi/resources/views/features/{$feature_name}/html/{$filename}.blade.php";
+            $res = file_put_contents($filepath, "");
+            if(!$res)
+            {
+                $result = ['error_code' => '0', 'error_msg' => 'SUCCESS', 'result' => [] ];
+            }
+        }
+        else
+        {
+            $filepath = "/var/www/xcloudserverapi/resources/views/features/{$feature_name}/js/{$filename}.blade.php";
+            $res = file_put_contents($filepath, "");
+            if(!$res)
+            {
+                $result = ['error_code' => '0', 'error_msg' => 'SUCCESS', 'result' => [] ];
+            }
+        }
+        return $result;
+    }
+
+    public function readDirectory($feature_name, $editmode)
+    {
+        $result = ['error_code' => 'FILEREADERROR', 'error_msg' => 'Unable to read directory', 'result' => [] ];
+        $filepath = "/var/www/xcloudserverapi/resources/views/features/{$feature_name}/{$editmode}";
+        $files = [];
+        $filenames = [];
+        if($handle = opendir($filepath))
+        {
+            while (false !== ($entry = readdir($handle))) {
+                $files[] = $entry;
+            }
+            // return $files;
+            foreach($files as $file)
+            {
+                $x = explode(".", $file);
+                if($x[0]!= "")
+                {
+                    $filenames[] = $x[0];
+                }
+            }
+            $result = ['error_code' => '0', 'error_msg' => 'SUCCESS', 'result' => [$filenames] ];
+        }
+        return $result;
+    }
+
+    public function saveCodeFile($param)
+    {
+        $result = ['error_code' => 'FILEWRITEERROR', 'error_msg' => 'Unable to write file', 'result' => [$param] ];
+        $filepath = "/var/www/xcloudserverapi/resources/views/features/{$param['feature']}/{$param['editmode']}/{$param['filename']}.blade.php";
+        $res = file_put_contents($filepath, $param['body']);
+        if($res)
+        {
+            $result = ['error_code' => '0', 'error_msg' => 'SUCCESS', 'result' => [] ];
+        }
+        return $result;
+    }
+
+    public function getFileCode($param)
+    {
+        $result = [];
+        $filepath = "/var/www/xcloudserverapi/resources/views/features/{$param['feature']}/{$param['editmode']}/{$param['filename']}.blade.php";
+        try
+        {
+            $res = file_get_contents($filepath);
+            $result = ['error_code' => '0', 'error_msg' => 'SUCCESS', 'result' => [$res] ];
+        }
+        catch(Exception $e)
+        {
+            $result = ['error_code' => 'FILEREADERROR', 'error_msg' => 'Unable to read file', 'result' => [$param] ];
+        }
+        return $result;
     }
 
 }
